backend/requirements.txt
# FastAPI + Server
fastapi[standard]>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.14.0,<2.0.0
aiosqlite>=0.20.0,<1.0.0
asyncpg>=0.30.0,<1.0.0

# Configuration
pydantic-settings>=2.7.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0

# HTTP & Crawling
httpx>=0.28.0,<1.0.0
playwright>=1.49.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
trafilatura>=2.0.0,<3.0.0
feedparser>=6.0.0,<7.0.0
robotexclusionrulesparser>=1.7.0,<2.0.0

# AI / LLM
google-genai>=1.5.0,<2.0.0
sentence-transformers>=3.4.0,<4.0.0
scikit-learn>=1.6.0,<2.0.0

# PDF Generation
weasyprint>=63.0,<70.0
jinja2>=3.1.0,<4.0.0

# Email
sendgrid>=6.11.0,<7.0.0

# Scheduling
apscheduler>=3.10.0,<4.0.0

# Observability
structlog>=24.4.0,<25.0.0

# Testing
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.25.0,<1.0.0
